+++ Shop/Models/DBModels/BillDetail.cs
@@ -8,9 +8,9 @@ namespace Shop.Models.DBModels
         public int IdBillDetail { get; set; }
         public int IdProduct { get; set; }
         public int Count { get; set; }
-        public double Discount { get; set; }
-        public double TotalPrice { get; set; }
-        public double LastPrice { get; set; }
+        public decimal Discount { get; set; }
+        public decimal TotalPrice { get; set; }
+        public decimal LastPrice { get; set; }
         public int IdBill { get; set; }
 
         public virtual Bill IdBillNavigation { get; set; }
